Make BOOST the default hash map backend and fix pycolmap build
- Default COLMAP_HASH_MAP_BACKEND to BOOST (faster boost::unordered maps);
  add boost-unordered as a core vcpkg dependency (needed by every build now)
  and drop the redundant hashmap-boost feature.
- FindDependencies.cmake: default the backend to STD when unset so downstream
  find_package(colmap) consumers (e.g. pycolmap) don't hit a fatal error; the
  exported colmap targets already carry the COLMAP_HASH_* macro they were built
  with.
- pycolmap: make the pybind11 type casters for the boost hash aliases visible in
  every binding TU. Flat map/set casters go in pybind11_extension.h (no flat
  type is opaque, so no collision); the NodeHashMap caster stays in
  scene/types.h next to the PYBIND11_MAKE_OPAQUE element-store maps so the opaque
  specializations always win. Add the missing includes to incremental_triangulator
  and correspondence_graph bindings.

Verified: colmap + pycolmap build under BOOST, and the full Python test suite
passes (879 passed, 0 failed); generated .pyi stubs no longer leak raw C++
container type names.
